Move database related stuff to db folder
