--- back-end/src/main/java/com/apellikka/fotosoppi/foto_soppi/ImageController.java
@@ -1,5 +1,0 @@
-package com.apellikka.fotosoppi.foto_soppi;
-
-public class ImageController {
-    
-}
