+++ back-end/src/main/java/com/apellikka/fotosoppi/foto_soppi/database/Image.java
@@ -1,4 +1,4 @@
-package com.apellikka.fotosoppi.foto_soppi;
+package com.apellikka.fotosoppi.foto_soppi.database;
 
 import jakarta.persistence.Entity;
