+++ back-end/src/main/java/com/apellikka/fotosoppi/foto_soppi/database/ImageController.java
@@ -1,0 +1,5 @@
+package com.apellikka.fotosoppi.foto_soppi.database;
+
+public class ImageController {
+    
+}
